Forsøk på å få inn startpunkt

+++ sandvika-activity-board-backend/src/main/java/no/bouvet/sandvika/activityboard/repository/ActivityRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import no.bouvet.sandvika.activityboard.domain.Activity;
 
 @RepositoryRestResource(collectionResourceRel = "activity", path = "activity")
-public interface ActivityRepository extends MongoRepository<Activity, Integer>
+public interface ActivityRepository extends MongoRepository<Activity, Long>
 {
     public List<Activity> findAllByBadgesIsNotNull();
 
@@ -20,7 +20,7 @@ public interface ActivityRepository extends MongoRepository<Activity, Integer>
 
     public List<Activity> findByStartDateLocalBetween(Date startDate, Date endDate);
 
-    public List<Activity> findById(int id);
+    public List<Activity> findById(long id);
 
     public List<Activity> findByStartDateLocalAfterAndType(Date startDate, String type);
